Deploy/requirements.txt
streamlit
numpy
cv2
keras
tensorflow
pillow